refactor: extract shared MidiAppClient to deduplicate MIDI port management
- Create src/midi/appClient.ts as shared client for MIDI IN/OUT ports
- Refactor controlMidiSender.ts to delegate to shared client (same API)
- Update midiBridge.ts to use shared markAppOutgoingAndForward()
- Eliminate duplication between service and driver (ports, shadow updates, setpoints)
- Keep MidiBridgeDriver focused on page-scoped passthrough orchestration
- Maintain existing Router.handleControl() API for controls.*.midi
- All tests pass (78/78)
