+++ config.yaml
@@ -104,28 +104,28 @@ pages:
       fader6:
         app: "voicemeeter"
         midi: {type: "passthrough"}
-      # mute1:
-      #   app: "voicemeeter"
-      #   midi: {type: "passthrough"}
-      # mute2:
-      #   app: "voicemeeter"
-      #   midi: {type: "passthrough"}
-      # mute3:
-      #   app: "voicemeeter"
-      #   midi: {type: "passthrough"}
-    passthroughs:
-      - driver: "midi"
-        to_port: "xtouch-gw"
-        from_port: "xtouch-gw-feedback"
-        filter:
-          channels: [1,2,3,4,5,6,7]
-          types: ["noteOn","noteOff","controlChange"] #,"pitchBend"]
-      - driver: "midi"
-        to_port: "qlc-in"
-        from_port: "qlc-out"
-        filter:
-          channels: [8,9]
-          types: ["noteOn","noteOff","controlChange","pitchBend"]
+      mute1:
+        app: "voicemeeter"
+        midi: {type: "passthrough"}
+      mute2:
+        app: "voicemeeter"
+        midi: {type: "passthrough"}
+      mute3:
+        app: "voicemeeter"
+        midi: {type: "passthrough"}
+    # passthroughs:
+    #   - driver: "midi"
+    #     to_port: "xtouch-gw"
+    #     from_port: "xtouch-gw-feedback"
+    #     filter:
+    #       channels: [1,2,3,4,5,6,7]
+    #       types: ["noteOn","noteOff","controlChange"] #,"pitchBend"]
+    #   - driver: "midi"
+    #     to_port: "qlc-in"
+    #     from_port: "qlc-out"
+    #     filter:
+    #       channels: [8,9]
+    #       types: ["noteOn","noteOff","controlChange","pitchBend"]
   - name: "Lum Latéraux"
     lcd:
       labels:
